LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/FileAPI/url/multi-global-origin-serialization.sub-expected.txt

Blocked access to external URL http://www1.localhost:8800/FileAPI/support/document-domain-setter.sub.html
Blocked access to external URL http://www2.localhost:8800/FileAPI/support/document-domain-setter.sub.html
Blocked access to external URL http://xn--lve-6lad.localhost:8800/FileAPI/support/document-domain-setter.sub.html
CONSOLE MESSAGE: SecurityError: Attempted to use a non-registrable domain.
CONSOLE MESSAGE: SecurityError: Blocked a frame with origin "http://localhost:8800" from accessing a cross-origin frame. Protocols, domains, and ports must match.


FAIL Blob URL serialization (specifically the origin) in multi-global situations SecurityError: Blocked a frame with origin "http://localhost:8800" from accessing a cross-origin frame. Protocols, domains, and ports must match.
